Picasso: Rename "smu.rs" to "mp1.rs".
